Develop >>> Main (#120)
* Analysisnamingcopy (#111)

* factor: update analysis naming screen copy

* factor: update analysis naming screen copy style

* Prtestbuild [ENH] Add a GitHub action to test .exe upon pull requests #113 (#118)

* refactor: separate building exe and creating new release

* upload artifacts only if triggered from release

* debug: handle workflow inputs

* remove precommit test

* test windows exe run on PR

* debug: fix executable filepath

* debug: fix filepath to updated name

* debug: path / to \

* refactor: implement run checks for macos

* debug: is_release == false instead of !

* Issue #110 new dev workflow documentation (#116)

* Add devops contributor workflow documentation

* Update release section

* feat: 🎸 basic excel importer (#112)

* feat: 🎸 basic excel importer

This Excel importer is very rudimentary but it gets the job done in most
cases. It assumes that the data to be imported lives in one of the
sheets in a "reasonable" position (i.e. headers in Row 1 and first
column is A1). If a file has multiple sheet, the user can change the
sheet to import. No explicit range selection in the sheet itself is
possible.

* run isort for styling

---------

Co-authored-by: DeanEby <ebyd21@gmail.com>

* Opendirectoryscreen copyupdate (#115)

* factor: remove initial output analysis

* refactor: context dependent post-analysis menu

* remove debug print statement

* refactor: slight copy changes

* debug: remove export outputs debug print

* test: 💍 expose CsvConfig for CsvTestData (#114)

* test: 💍 expose CsvConfig for CsvTestData

The CsvConfig class encodes how csv files can be loaded for testing,
but it was never exposed or wired up in the CsvTestData file. This
commit corrects this, and adds an example of how to use it to customize
the way test CSV files are loaded for testing.

* style: 💄 isort && black

---------

Co-authored-by: soul-codes <40335030+soul-codes@users.noreply.github.com>